backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.3.0
requests==2.32.3
ultralytics==8.2.0
